added empty tests and refactored code
