+++ nistats/thresholding.py
@@ -7,6 +7,8 @@ import numpy as np
 from scipy.ndimage import label
 from scipy.stats import norm
 from nilearn.input_data import NiftiMasker
+from nilearn.image.resampling import coord_transform
+import pandas as pd
 
 
 def fdr_threshold(z_vals, alpha):
@@ -48,7 +50,7 @@ def map_threshold(stat_img, mask_img=None, threshold=.001,
     -------
     thresholded_map : Nifti1Image,
         the stat_map theresholded at the prescribed voxel- and cluster-level
-        
+
     threshold: float,
         the voxel-level threshold used actually
     """
@@ -83,3 +85,69 @@ def map_threshold(stat_img, mask_img=None, threshold=.001,
             stats[labels == label_] = 0
 
     return masker.inverse_transform(stats), z_th
+
+
+def get_clusters_table(stat_img, stat_threshold, cluster_threshold):
+    """Creates pandas dataframe with img cluster statistics.
+
+    Parameters
+    ----------
+    stat_img : Niimg-like object,
+       statistical image (presumably in z scale)
+
+    stat_threshold: float, optional
+        cluster forming threshold (either a p-value or z-scale value)
+
+    cluster_threshold : float, optional
+        cluster size threshold
+
+    Returns
+    -------
+    Pandas dataframe with img clusters
+    """
+
+    stat_map = stat_img.get_data()
+
+    # Extract connected components above threshold
+    label_map, n_labels = label(stat_map > stat_threshold)
+
+    # labels = label_map[search_mask.get_data() > 0]
+    for label_ in range(1, n_labels + 1):
+        if np.sum(label_map == label_) < cluster_threshold:
+            stat_map[label_map == label_] = 0
+
+    label_map, n_labels = label(stat_map > stat_threshold)
+    label_map = np.ravel(label_map)
+    stat_map = np.ravel(stat_map)
+
+    peaks = []
+    max_stat = []
+    clusters_size = []
+    coords = []
+    for label_ in range(1, n_labels + 1):
+        cluster = stat_map.copy()
+        cluster[label_map != label_] = 0
+
+        peak = np.unravel_index(np.argmax(cluster),
+                                stat_img.get_data().shape)
+        peaks.append(peak)
+
+        max_stat.append(np.max(cluster))
+
+        clusters_size.append(np.sum(label_map == label_))
+
+        x_map, y_map, z_map = peak
+        mni_coords = np.asarray(
+            coord_transform(x_map, y_map, z_map, stat_img.affine)).tolist()
+        mni_coords = [round(x) for x in mni_coords]
+        coords.append(mni_coords)
+
+    vx, vy, vz = zip(*peaks)
+    x, y, z = zip(*coords)
+
+    columns = ['Vx', 'Vy', 'Vz', 'X', 'Y', 'Z', 'Peak stat', 'Cluster size']
+    clusters_table = pd.DataFrame(
+        list(zip(vx, vy, vz, x, y, z, max_stat, clusters_size)),
+        columns=columns)
+
+    return clusters_table
