+++ nistats/utils.py
@@ -347,6 +347,10 @@ def get_design_from_fslmat(fsl_design_matrix_path, column_names=None):
     """ Extract design matrix dataframe from FSL mat file.
     """
     design_matrix_file = open(fsl_design_matrix_path, 'r')
+    # Based on the openneuro example this seems to be the right
+    # marker to start extracting the matrix until the end of the file
+    # Conventions of FSL mat files should be verified in more detail for
+    # a general case
     for line in design_matrix_file:
         if '/Matrix' in line:
             break
